requirements.txt
python-decouple==3.8
termcolor==1.1.0
